refactor: update up and cd, rename vars
