+++ src/nwd/cd.js
@@ -1,47 +1,32 @@
-import * as path from 'path';
-import { goUp } from './up.js';
-import { checkDirentExist } from './checkDirentExist.js';
+import { homedir } from 'os';
+import { getAbsolutePath } from '../utils/getAbsolutePath.js';
+import { checkDirentExist } from '../utils/checkDirentExist.js';
 import { stat } from 'fs/promises';
 import { ERROR_MESSAGE } from '../constants/errorMessage.js';
 
-export const goToFolder = async (currentDir, dirPath) => {
-  const rootDir = path.parse(process.cwd()).root;
+export const goToDir = async (currentDir, direntPath) => {
   let newDir;
 
   try {
-    if (!dirPath) {
-      throw new Error(`${ERROR_MESSAGE}: no path! Enter command in format cd path_to_directory!`)
-    }
-    
-    if (dirPath.startsWith(rootDir)) {
-      newDir = dirPath;
+    if (!direntPath || direntPath === '~') {
+      newDir = homedir();
     } else {
-      newDir = dirPath.split('/').reduce((acc, folder) => {
-        if (folder && folder === '..') {
-          acc = goUp(acc);
-        } else if (folder && folder !== '.') {
-          acc = acc === rootDir ? `${acc}${folder}` : `${acc}/${folder}`;
-        } 
-        return acc;
-      }, currentDir);
-    }
+      newDir = getAbsolutePath(currentDir, direntPath);
 
-    const isDirentExist = await checkDirentExist(newDir);
+      const isDirentExist = await checkDirentExist(newDir);
 
-    if (!isDirentExist) {
-      newDir = currentDir;
-      throw new Error(`${ERROR_MESSAGE}: directory doesn't exist!`);
-    } else {
-      const newDirStat = await stat(newDir);
-      
-      if (!newDirStat.isDirectory()) {
+      if (!isDirentExist) {
         newDir = currentDir;
-        throw new Error(`${ERROR_MESSAGE}: not a directory!`);
+        throw new Error(`${ERROR_MESSAGE}: directory doesn't exist!`);
       } else {
-        newDir = path.resolve(newDir);
+        const newDirStat = await stat(newDir);
+      
+        if (!newDirStat.isDirectory()) {
+          newDir = currentDir;
+          throw new Error(`${ERROR_MESSAGE}: not a directory!`);
+        }
       }
     }
-   
   } catch (error) {
     console.error(error.message);
   }
