+++ src/nwd/up.js
@@ -1,10 +1,8 @@
-import * as path from 'path';
+import { parse, normalize } from 'path';
 
 export const goUp = (currentDir) => {
-  const rootDir = path.parse(process.cwd()).root;
-  const newDir = (currentDir !== rootDir) && (currentDir.indexOf('/') !== currentDir.lastIndexOf('/'))
-    ? currentDir.slice(0, currentDir.lastIndexOf('/'))
-    : rootDir;
+  const { root, dir } = parse(currentDir);
+  const newDir = normalize(`${root}${dir}`);
 
   return newDir;
 };
