+++ src/utils/checkDirentExist.js
